+++ lib/Extension/LanguageServerCodeTransform/LanguageServerCodeTransformExtension.php
@@ -274,6 +274,16 @@ class LanguageServerCodeTransformExtension implements Extension
             LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
         ]);
 
+        $container->register(TransformerCodeActionPovider::class.'add_missing_return_types', function (Container $container) {
+            return new TransformerCodeActionPovider(
+                $container->get('code_transform.transformers'),
+                'add_missing_return_types',
+                'Add missing return types'
+            );
+        }, [
+            LanguageServerExtension::TAG_DIAGNOSTICS_PROVIDER => [],
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+        ]);
 
         $container->register(GenerateMethodProvider::class, function (Container $container) {
             return new GenerateMethodProvider(