+++ lib/WorseReflection/Core/Type/ClassNamedType.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Type;
+
+use Phpactor\WorseReflection\Core\ClassName;
+
+interface ClassNamedType
+{
+    public function name(): ClassName;
+}