+++ lib/WorseReflection/Core/Type/ClassType.php
@@ -6,7 +6,7 @@ use Phpactor\WorseReflection\Core\ClassName;
 use Phpactor\WorseReflection\Core\Trinary;
 use Phpactor\WorseReflection\Core\Type;
 
-class ClassType extends Type
+class ClassType extends Type implements ClassNamedType
 {
     public ClassName $name;