+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Improvements:
 
 Features:
 
+  - [ct] Add missing PHP return types
   - [wr] Support for inference for `array_map`, and arrow and anonymous functions
   - [ct] Add missing @return type docblocks code transformation
   - [cmp] Explicitly enable/disable completors and disable `keyword` completor by default.